Release 0.9.0 (2023-04-10) (#36) (#37)
### Features

* Add a total realtime QPS metric
* Add aws supplied e2e latency
* Add basic UDF functional tests for v2
* Add error counters for realtime updates
* Add functional test stubs for v2
* Add target to generate delta for sample udf.js
* Add test data artifacts to dist/test_data/deltas
* Add UDF delta file generator tool.
* Add UDF delta file upload through terraform config.
* Add udf.js delta file to test set
* Upgrade to build-system 0.22.0 and functionaltest-system 0.2.0

### Bug Fixes

* Add a dashboard for environments
* Add documentation for editing dashboards
* Change envoy log level to debug
* Check that recovery function is valid before calling it.
* Enable docker network cleanup
* Ensure changelog notes use specific version
* ignore interestGroupNames argument
* MetricsRecorder no longer a singleton.
* MetricsRecorder now optional for retry templates.
* Return missing key error status from internal lookup server.
* Upgrade gRPC and make lookup client a singleton
* Use dynamic_cast to get metric_sdk::MeterProvider provider.

### Terraform

* Add us-west-1 terraform

### Documentation

* Add documentation on roma child processes.
* Add instructions on realtime updates
* Add note to use grpcurl for v2 since http has a bug.
* Add v2 request JSON schema
* AWS realtime update capabilities
* Correct udf target name
* Update documentation for building data cli.
* Update realtime metrics querying docs

Bug: 277075104
GitOrigin-RevId: 755a69bbf847c152f2e3e4ddf0f42e6410884144

+++ components/aws/sqs_cleanup_manager.py
@@ -124,6 +124,75 @@ class SqsCleanupManager(object):
                     targets_to_subscriptions[target_arn] = subscription_arn
         return targets_to_subscriptions
 
+    def _get_all_subscriptions(self) -> dict:
+        """Find all subscriptions.
+
+        Returns:
+        Dictionary of arns, from subscription target (queue) to subscription.
+        """
+        client = self._sns_resource.meta.client
+        paginator = client.get_paginator("list_subscriptions")
+        targets_to_subscriptions = {}
+        for response in paginator.paginate():
+            for subscription in response.get("Subscriptions", []):
+                subscription_arn = subscription.get("SubscriptionArn")
+                target_arn = subscription.get("Endpoint")
+                if subscription is not None and target_arn is not None:
+                    targets_to_subscriptions[target_arn] = subscription_arn
+        return targets_to_subscriptions
+
+    def _does_queue_exist(self, queue_arn: str) -> bool:
+        """Check if a queue exists.
+
+        Args:
+          queue_arn: SQS ARN string.
+
+        Returns:
+         bool indicating if the SQS exists.
+        """
+        # aws does not provide a better way to get a name from an ARN
+        queue_name = queue_arn.split(":")[-1]
+        sqsclient = self._sqs_resource.meta.client
+        # aws does not provide a better way to check if an SQS exists
+        try:
+            sqsclient.get_queue_url(QueueName=queue_name)
+        except:
+            return False
+
+        return True
+
+    def _get_dangling_subscriptions(self, topic_arn: str) -> list:
+        """Find all subscriptions with missing queues.
+
+        Args:
+          topic_arn: SQS topic ARN string.
+
+        Returns:
+        List of subscription ARNs that have a missing SQS.
+        """
+        client = self._sns_resource.meta.client
+        paginator = client.get_paginator("list_subscriptions_by_topic")
+        dangling_subscriptions = []
+        for response in paginator.paginate(TopicArn=topic_arn):
+            for subscription in response.get("Subscriptions", []):
+                subscription_arn = subscription.get("SubscriptionArn")
+                target_arn = subscription.get("Endpoint")
+                if self._does_queue_exist(target_arn):
+                    continue
+
+                dangling_subscriptions.append(subscription_arn)
+        return dangling_subscriptions
+
+    def _delete_subscriptions(self, subscriptions_to_delete: list) -> None:
+        """Delete subscriptions.
+
+        Args:
+          subscriptions_to_delete: a list of subscriptions to delete.
+        """
+        for subscription_arn in subscriptions_to_delete:
+            subscription = self._sns_resource.Subscription(subscription_arn)
+            subscription.delete()
+
     def _cleanup(
         self,
         expired_queues: list,
@@ -146,8 +215,31 @@ class SqsCleanupManager(object):
                 subscription = self._sns_resource.Subscription(subscription_arn)
                 subscription.delete()
                 subscriptions_deleted += 1
-            queue.delete()
-        return len(expired_queues), subscriptions_deleted
+                queue.delete()
+        return subscriptions_deleted, subscriptions_deleted
+
+    def _cleanup_orphan_queues(
+        self,
+        expired_queues: list,
+        subscriptions: Dict[str, str],
+    ) -> int:
+        """Deletes all expired SQS queues missing an SNS.
+
+        Args:
+          expired_queues: List of expired SQS queues.
+          subscriptions: Subscription arns to subscription.
+
+        Returns:
+          number of queues deleted.
+        """
+        queues_deleted = 0
+        for queue in expired_queues:
+            queue_arn = queue.attributes.get("QueueArn")
+            subscription_arn = subscriptions.get(queue_arn)
+            if subscription_arn is None:
+                queue.delete()
+                queues_deleted += 1
+        return queues_deleted
 
     def find_and_cleanup(
         self,
@@ -155,11 +247,16 @@ class SqsCleanupManager(object):
         queue_prefix: str,
         timeout_secs: int,
     ) -> Tuple[int, int]:
-        """Deletes all expired Queues and associated subscriptions.
+        """Deletes all expired Queues and associated subscriptions. Deletes dangling subscriptions.
 
         Returns:
           Number of deleted queues and subscriptions.
         """
-        target_subscriptions = self._get_target_subscriptions(topic_arn)
+        all_subscriptions = self._get_all_subscriptions()
         expired_queues = self._find_expired_queues(queue_prefix, timeout_secs)
+        self._cleanup_orphan_queues(expired_queues, all_subscriptions)
+        dangling_subscriptions = self._get_dangling_subscriptions(topic_arn)
+        self._delete_subscriptions(dangling_subscriptions)
+        expired_queues = self._find_expired_queues(queue_prefix, timeout_secs)
+        target_subscriptions = self._get_target_subscriptions(topic_arn)
         return self._cleanup(expired_queues, target_subscriptions)
